Fixing interface methods test

+++ dcevm/src/test/java8/com/github/dcevm/test/methods/InterfaceMethodsTest.java
@@ -56,7 +56,7 @@ public class InterfaceMethodsTest {
 
   // Version 1
   public static interface A___1 {
-    public static long staticMethod() {
+    public static int staticMethod() {
       return 2;
     }
   }
